fix grpc query server not setting max inbound msg size
